Import Volatility script
This script automates volatility tasks.

Mainly, it allows you to set profile and source file by outputing them
in .profile and .memdump files. The script can then be used to run
plugins without having to specify those informations which saves a
lot of time, place on the command line and allows for quick return to a case